third_party/WebKit/LayoutTests/fullscreen/full-screen-popup.html

<body>
<script src="full-screen-test.js"></script>
<span></span>
<script>
    // Bail out early if the full screen API is not enabled or is missing:
    if (Element.prototype.webkitRequestFullScreen == undefined) {
        logResult(false, "Element.prototype.webkitRequestFullScreen == undefined");
        endTest();
    } else {
        if (window.testRunner) {
            testRunner.setCanOpenWindows();
            testRunner.setPopupBlockingEnabled(true);
            testRunner.setCloseRemainingWindowsWhenComplete(true);
        }
        var span = document.getElementsByTagName('span')[0];

        var fullScreenAndPopup = function() {
            span.webkitRequestFullScreen();
            testExpected("window.open('about:blank')", null);
            document.webkitCancelFullScreen();
            setTimeout(endTest, 0);
        };

        runWithKeyDown(fullScreenAndPopup);
    }
</script>
