third_party/WebKit/LayoutTests/fullscreen/full-screen-popup-expected.txt

EXPECTED (window.open('about:blank') == 'null') OK
END OF TEST
